requirements.txt
Pillow==7.2.0
numpy==1.16.2
scipy==1.3.1
